pyproject.toml
[project]
name = "teste-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "airflow-operators>=0.11.0",
    "apache-airflow>=3.1.7",
    "kagglehub[pandas-datasets]>=1.0.0",
    "pathlib>=1.0.1",
]
